[FIX] TestTime 타입을 Long으로 통일

+++ src/main/java/swm_nm/morandi/test/domain/Test.java
@@ -20,7 +20,7 @@ public class Test {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long testId;
     private LocalDateTime testDate; // 테스트 시작 시간
-    private Integer testTime; // minutes
+    private Long testTime; // minutes
     private Integer problemCount;
     private Long remainingTime; // seconds
